% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Cancer and height}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1250 rows and 1 columns.}
\usage{
cancer
}
\description{
Cancer and height
}
\keyword{datasets}
