% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curry.R
\docType{data}
\name{curry}
\alias{curry}
\title{Curry long shots}
\format{
A tbl_df with with 27 obs on 1 variables
\describe{
  \item{shot}{the result of a basket attempt, either "hit" or "miss"}
 }
}
\usage{
curry
}
\description{
In the 2016 season, it was noted that professional basketball player Steph Curry remarkable basket-shooting performance beyond 30 feet.  This data frame contains his long range shooting performance across 27 attempts. For reference the long range shooting percentage of NBA players that season was 7.5%.
}
\keyword{datasets}
