% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfood1.R
\docType{data}
\name{fastfood1}
\alias{fastfood1}
\title{Card & Krueger minimum wage data, cross-sectional}
\format{

}
\usage{
fastfood1
}
\description{
Cross-sectional excerpt of employment data collected from New Jersey and
Pennsylvania fast food restaurants as part of Card & Krueger's seminal study
on minimum wage.  This version of the dataset retains covariate information
from the first wave of data collection only and employment numbers from the
second wave of data collection only (see fastfood2 for a different
version of the dataset with repeated measures). Of the 410 restaurants
originally surveyed, 11 have been excluded here because they were closed
or refused to participate during the second wave of data collection.  Missing
values have also been imputed.
}
\keyword{datasets}
