% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfood2.R
\docType{data}
\name{fastfood2}
\alias{fastfood2}
\title{Card & Krueger minimum wage data, panel version}
\format{

}
\usage{
fastfood2
}
\description{
Panel employment data collected from New Jersey and
Pennsylvania fast food restaurants as part of Card & Krueger's seminal study
on minimum wage.  This version of the dataset retains covariate and outcome
measures from both waves of data collection (see fastfood1 for a cross-
sectional version of the dataset). Of the 410 restaurants
originally surveyed, 11 have been excluded here because they were closed
or refused to participate during the second wave of data collection.  Missing
values have also been imputed.
}
\keyword{datasets}
