% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first.r
\name{get_first}
\alias{get_first}
\title{Get first digit}
\usage{
get_first(x)
}
\arguments{
\item{x}{a large number (of class numeric or integer)}
}
\value{
The first digit.
}
\description{
Returns the first digit of a larger number.
}
