% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_job.R
\docType{data}
\name{grad_job}
\alias{grad_job}
\title{Job prospects for Cal graduates}
\format{
A data.frame with 20 rows and 5 variables:
\describe{
  \item{GPA}{GPA of the student.}
  \item{years_exp}{Years of job experience.}
  \item{rec}{Categorical variable denoting strength of the student's
  recommendation letters, with levels "strong", "moderate", and "weak".}
  \item{cal_grad}{Logical variable indicating whether the student is a Cal
  graduate.}
  \item{good_job}{Categorical avariable denoting whether the student holds a
  good job.}
  }
}
\usage{
grad_job
}
\description{
Small synthetic dataset containing data on hypothetical students, some of
whom are Cal graduates, and information about their employment status.
}
\keyword{datasets}
