% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung.R
\docType{data}
\name{lung}
\alias{lung}
\title{Lung Disease and Therapy}
\format{A tbl_df with 146 rows and 2 columns:
\describe{
  \item{IQ}{IQ score}
  \item{group}{treatment group to which the subject was assigned, either \code{dexamethasone} or \code{placebo}}.
  }}
\usage{
lung
}
\description{
Lung Disease and Therapy
}
\keyword{datasets}
