% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millenials.R
\docType{data}
\name{millenials}
\alias{millenials}
\title{Millenials and Marriage}
\format{A tbl_df with with 75 obs on 2 variables
\describe{
  \item{id}{a subject identifier}
  \item{response}{response to the question: "Do you favor same-sex marriage?"}
 }}
\usage{
millenials
}
\description{
Results from a 2014 Pew poll of a random sample of 75 millenials (Americans born after 1980).
}
\keyword{datasets}
