% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millennials.R
\docType{data}
\name{millennials}
\alias{millennials}
\title{Millennials and Marriage}
\format{
A tbl_df with with 75 obs on 2 variables
\describe{
  \item{id}{a subject identifier}
  \item{response}{response to the question: "Do you favor same-sex marriage?"}
 }
}
\usage{
millennials
}
\description{
Results from a 2014 Pew poll of a random sample of 75 millennials (Americans born after 1980).
}
\keyword{datasets}
