% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nypd.R
\docType{data}
\name{nypd}
\alias{nypd}
\title{NYPD}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 902 rows and 2 columns.}
\usage{
nypd
}
\description{
Data concerning stop and frisk policies conducted by Quinnipiac 2013.
}
\keyword{datasets}
