% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote.R
\docType{data}
\name{promote}
\alias{promote}
\title{Results from study on sex and promotion decisions}
\format{A tbl_df with with 48 obs on 3 variables
\describe{
  \item{id}{a subject identifier}
  \item{sex}{The sex commonly associated with the name on the application}
  \item{decision}{the decision of the bank manager to promote or not promote}
 }}
\usage{
promote
}
\description{
Results from study on sex and promotion decisions
}
\keyword{datasets}
