% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey141.R
\docType{data}
\name{survey141}
\alias{survey141}
\title{Cleaned Student Survey Data}
\format{A tbl_df with 53 rows and 33 variables:
\describe{
  \item{social}{Personal political views on social issues, 
  with 1 being extremely liberal and 10 being extremely conservative}
  \item{economic}{Personal Political views on economic issues}
  \item{diet}{Dietary habits of respondent}
  \item{college_app}{Number of colleges/universities applied to}
  \item{reedie_social}{Respondent estimate of average Reedie's views on social issues}
  \item{reedie_economic}{Estimate of average Reedie's views on economic issues}
  \item{study}{Best study spot on campus}
  \item{year}{Class standing}
  \item{commons}{Best station at Commons}
  \item{transportation}{Most used transportation method}
  \item{division}{Division of (intended) major}
  \item{tradition}{Best Reed thing/tradition}
  \item{awkward}{Response to "Reedies are socially awkward."}
  \item{technology}{Response to "Technology is destructive to interpersonal relationships."}
  \item{historian}{Respondent's favorite ancient Greek historian}
  \item{alcohol}{How often respondent drinks alcohol a week on average}
  \item{reedie_alcohol}{Estimate for how often the average Reedie drinks alcohol per week on average}
  \item{marijuana}{How often respondent uses marijuana on average}
  \item{reedie_marijuana}{Estimate for how often the average Reedie uses marijuana per week on average}
  \item{social_media}{Social media used by respondent}
  \item{coffee_tea}{Respondent's drink of choice}
  \item{computer}{Respondent's computer of choice}
  \item{season}{Respondent's favorite season}
  \item{thai}{Best Thai restaurant}
  \item{ac}{Does respondent's house/apartment/dorm have AC}
  \item{hogwarts}{Hogwarts house identification}
  \item{beach_mountain}{Beach or mountains?}
  \item{donut}{Best donut shop}
  \item{first_kiss}{Respondent's age at first kiss}
  \item{meme}{Favorite meme}
  \item{dog_pants}{How a dog would wear pants}
  \item{hot_dog}{Sandwich status of hod dogs}
  \item{duck_horse}{Who the respondent would rather fight}
  }}
\usage{
survey141
}
\description{
Survey results from Reed Math 141 students, Spring 2018. Each row
represents a single respondent. The data set has been wrangled to make
generating plots easier.
}
\keyword{datasets}
