% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unvotes.R
\docType{data}
\name{unvotes}
\alias{unvotes}
\title{Votes at the United Nations}
\format{
A tbl_df with with 419966 obs on 15 variables, three of which include...
\describe{
  \item{rcid}{the roll call id}
  \item{country}{the country name}
  \item{year}{the year of the vote}
  \item{issue}{the general issue of the resolution}
 }
}
\usage{
unvotes
}
\description{
Country-level voting results on every UN resolution concerning three issues: Colonialism,
Human rights, and Nuclear weapons from 1946 - 2019.
}
\keyword{datasets}
