% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yawn.R
\docType{data}
\name{yawn}
\alias{yawn}
\title{Results from MythBusters: Is Yawning Contagious?}
\format{
A tbl_df with with 50 obs on 3 variables
\describe{
  \item{id}{an identifier for each subject}
  \item{group}{the group assigned to each subject: stimulus (in presence of a yawn) and no stimulus (not in the presence of a yawn)}
  \item{response}{whether or not the subject yawned}
 }
}
\usage{
yawn
}
\description{
Results from MythBusters: Is Yawning Contagious?
}
\keyword{datasets}
