% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_survey.R
\docType{data}
\name{class_survey}
\alias{class_survey}
\title{Class Survey Data}
\format{
A tbl_df with with 619 rows and 31 variables:
\describe{

  \item{year}{number of years/semesters spent at UC Berkeley}
  \item{major}{declared or intended major}
  \item{coding_exp_words}{degree of coding experience on a phrased scale}
  \item{coding_exp_scale}{degree of coding experience on a scale of 1 to 10}
  \item{coding_languages}{programming languages used among those who have coding experience}
  \item{favorite_thing_cal}{favorite thing about attending UC Berkeley}
  \item{body_piercings}{number of body piercings}
  \item{pets_plus_siblings}{number of (combined) pets and siblings}
  \item{diet}{dietary practice in terms of eating plants/animals}
  \item{title}{most closely held academic/professional label}
  \item{study_place}{favorite place to study on the UC Berkeley campus}
  \item{olympic_sport}{favorite Winter Olympic sport}
  \item{season}{favorite season in the year}
  \item{beach_or_mtns}{preference for being at mountains or the beach}
  \item{coffee_or_tea}{preference for coffee or tea}
  \item{best_boba_shop}{favorite boba tea shop in Berkeley}
  \item{tech}{degree of agreement with the statement: "Technology is destructive to interpersonal relationships" on a scale of 1 to 10}
  \item{climate}{degree of agreement with the statement: "Humanity will be able to use its engineering prowess to counteract the effects of climate change" on a scale of 1 to 10}
  \item{crypto}{degree of agreement with the statement: "Cryptocurrency will play a dominant role in the global financial system" on a scale of 1 to 10}
  \item{new_COVID_variant}{stated probability that another COVID variant (other than Delta, Omicron) will disrupt class instruction in Fall 2022}
  \item{fire_alarm}{stated probability that a fire alarm will go off in one of the UC Berkeley dorms this week}
  \item{are_hotdogs_sandwiches}{answer to the question: "Are hotdogs sandwiches?"}
  \item{horse_duck}{answer to the question: "Who would you rather fight: a horse-sized duck or a duck-sized horse?"}
  \item{dog_pants}{answer to the question: "If a dog wore pants would he wear them on both pairs of legs or just his hind legs?" These choices corresponded to a left image and right image, respectively. }
  \item{artist}{\code{TRUE} if student identifies most accurately as an Artist, \code{FALSE} otherwise}
  \item{humanist}{\code{TRUE} if student identifies most accurately as an Humanist, \code{FALSE} otherwise}
  \item{natural_scientist}{\code{TRUE} if student identifies most accurately as a Natural Scientist, \code{FALSE} otherwise}
  \item{social_scientist}{\code{TRUE} if student identifies most accurately as a Social Scientist, \code{FALSE} otherwise}
  \item{comp_scientist}{\code{TRUE} if student identifies most accurately as a Computational Scientist, \code{FALSE} otherwise}
  \item{entrepreneur}{\code{TRUE} if student identifies most accurately as an Entrepreneur, \code{FALSE} otherwise}
}
}
\source{
The data are adapted and cleaned from a Google Form.
}
\usage{
class_survey
}
\description{
This data set contains responses to a class survey.

The class survey sent to students included questions regarding
academic status, current world events, COVID-19 variants, their personal lives
and whether one would fight a duck-sized horse or a horse-sized duck.
}
\keyword{datasets}
